refactor(blorktools): html2canvas loader
